refactor: normalize file name casing for better readability and convention
