+++ client/src/features/auth/pages/SignIn/SignInForm/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from "./use-sign-in-form";
+export * from "./useSignInForm";
