+++ client/src/features/auth/pages/SignIn/SignInForm/utils/index.ts
@@ -1,1 +1,1 @@
-export * from "./handle-sign-in";
+export * from "./handleSignIn";
